Rename `contents/` to `content/`
